LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-img-element/image-loading-lazy-in-cross-origin-iframe-002.sub-expected.txt



FAIL A loading=lazy image in a below-viewport cross-origin iframe loads only
       when the cross-origin iframe is scrolled into view assert_equals: The loading=lazy image should not block the iframe's load event expected "window_loaded" but got "image_loaded"
